Seeds widen from u32 to u64 through the engine, the Python surface, the
batch and WASM entry points and the spec and manifest; the random streams
derive the same state for every seed below 2**32, so every known answer and
record is unchanged, and tests/known_answer_seed64.json holds the digest of a
seed above it. The merge onto fix/ptv20-final's pt-v20 had no conflicts.
The pt-v20 repricing test helper that fix/ptv20-core added still takes a u32
seed against seed64's Engine::with_params; it is fixed in its own commit
after the sandbox merge, as on trial/0.8.5-integration (079cc6c).

+++ rust/examples/divergence.rs
@@ -168,14 +168,16 @@ pub fn build_company(c: &Json) -> TickCompany {
 /// differ in how they draw, never in where they start.
 struct Setup {
     engine: Engine,
-    seed: u32,
+    seed: u64,
     ticks: usize,
     volatility: f64,
 }
 
 fn setup(doc: &Json) -> Setup {
     let spec = &doc["spec"];
-    let seed = spec["tickSeed"].as_u64().unwrap() as u32;
+    // The reference's seeds are 32-bit (`>>> 0`), so the recorded seed is
+    // read at that width and widened for the engine, which takes a u64.
+    let seed = u64::from(spec["tickSeed"].as_u64().unwrap() as u32);
     let ticks = spec["ticks"].as_i64().unwrap() as usize;
     let volatility = bits(spec["volatility"].as_str().unwrap());
